+++ helpers/documentHelper.js
@@ -19,69 +19,40 @@ const requestHelper = require("./requestHelper.js");
 
 var documentHelper = {};
 
-const SUPPORTED_FORMATS = {
-    "djvu": {"type": "word"},
-    "doc": {"type": "word"},
-    "docm": {"type": "word"},
-    "docx": {"type": "word", "edit": true},
-    "dot": {"type": "word"},
-    "dotm": {"type": "word"},
-    "dotx": {"type": "word"},
-    "epub": {"type": "word"},
-    "fb2": {"type": "word"},
-    "fodt": {"type": "word"},
-    "html": {"type": "word"},
-    "mht": {"type": "word"},
-    "odt": {"type": "word"},
-    "ott": {"type": "word"},
-    "oxps": {"type": "word"},
-    "pdf": {"type": "word"},
-    "rtf": {"type": "word"},
-    "txt": {"type": "word"},
-    "xps": {"type": "word"},
-    "xml": {"type": "word"},
-
-    "csv": {"type": "cell"},
-    "fods": {"type": "cell"},
-    "ods": {"type": "cell"},
-    "ots": {"type": "cell"},
-    "xls": {"type": "cell"},
-    "xlsm": {"type": "cell"},
-    "xlsx": {"type": "cell", "edit": true},
-    "xlt": {"type": "cell"},
-    "xltm": {"type": "cell"},
-    "xltx": {"type": "cell"},
-
-    "fodp": {"type": "slide"},
-    "odp": {"type": "slide"},
-    "otp": {"type": "slide"},
-    "pot": {"type": "slide"},
-    "potm": {"type": "slide"},
-    "potx": {"type": "slide"},
-    "pps": {"type": "slide"},
-    "ppsm": {"type": "slide"},
-    "ppsx": {"type": "slide"},
-    "ppt": {"type": "slide"},
-    "pptm": {"type": "slide"},
-    "pptx": {"type": "slide", "edit": true}
-};
-
 documentHelper.getFileExtension = function (fileName) {
     var parts = fileName.toLowerCase().split(".");
 
     return parts.pop();
 }
 
-documentHelper.getDocumentType = function (extension) {
-    if (SUPPORTED_FORMATS[extension] !== undefined) {
-        return SUPPORTED_FORMATS[extension]["type"];
+documentHelper.getDocumentType = function (addon, extension) {
+    const formats = addon.config.docServer().formats || [];
+
+    for (let i = 0; i < formats.length; i++) {
+        if (formats[i].name === extension) return formats[i].type;
     }
 
     return null;
 }
 
-documentHelper.isEditable = function (extension) {
-    return SUPPORTED_FORMATS[extension] !== undefined && SUPPORTED_FORMATS[extension]["edit"] === true;
+documentHelper.isEditable = function (addon, extension) {
+    const formats = addon.config.docServer().formats || [];
+
+    for (let i = 0; i < formats.length; i++) {
+        if (formats[i].name === extension) return formats[i].actions.includes('edit');
+    }
+
+    return false;
+}
+
+documentHelper.isViewable = function (addon, extension) {
+    const formats = addon.config.docServer().formats || [];
+
+    for (let i = 0; i < formats.length; i++) {
+        if (formats[i].name === extension) return formats[i].actions.includes('view');
+    }
+
+    return false;
 }
 
 documentHelper.getDocumentKey = function (attachmentId, updateTime) {
@@ -97,7 +68,7 @@ documentHelper.getEditorConfig = async function (addon, httpClient, clientKey, l
 
     let permissionEdit = await requestHelper.checkContentPermission(httpClient, userInfo.accountId, attachmentInfo.id, "update");
 
-    if (permissionEdit && documentHelper.isEditable(fileType)) {
+    if (permissionEdit && documentHelper.isEditable(addon, fileType)) {
         mode = "edit";
         callbackUrl = await urlHelper.getCallbackUrl(addon, localBaseUrl, clientKey, userInfo.accountId, attachmentInfo.pageId || attachmentInfo.blogPostId, attachmentInfo.id);
     }
@@ -106,7 +77,7 @@ documentHelper.getEditorConfig = async function (addon, httpClient, clientKey, l
         type: "desktop",
         width: "100%",
         height: "100%",
-        documentType: documentHelper.getDocumentType(fileType),
+        documentType: documentHelper.getDocumentType(addon, fileType),
         document: {
             title: attachmentInfo.title,
             url: await urlHelper.getFileUrl(addon, localBaseUrl, clientKey, userInfo.accountId, attachmentInfo.pageId || attachmentInfo.blogPostId, attachmentInfo.id),